+++ Rdmp.Core/DataLoad/Modules/DataFlowOperations/ColumnForbidder.cs
@@ -27,11 +27,11 @@ public class ColumnForbidder : IPluginDataFlowComponent<DataTable>
 {
     private Regex _crashIfAnyColumnMatches;
 
-    [DemandsInitialization("Crashes the load if any column name matches this regex")]
+    [DemandsInitialization("Crashes the load if any column name matches this regex. This option takes precedence over any selected Standard Regex")]
     public Regex CrashIfAnyColumnMatches
     {
         get => _crashIfAnyColumnMatches;
-        set => _crashIfAnyColumnMatches = new Regex(value.ToString(), value.Options | RegexOptions.IgnoreCase);
+        set => _crashIfAnyColumnMatches = value != null?new Regex(value.ToString(), value.Options | RegexOptions.IgnoreCase):null;
     }
 
     [DemandsInitialization(
@@ -46,7 +46,7 @@ public class ColumnForbidder : IPluginDataFlowComponent<DataTable>
 
     public DataTable ProcessPipelineData(DataTable toProcess, IDataLoadEventListener listener,
         GracefulCancellationToken cancellationToken)
-    {
+    { 
         var checkPattern = _reCache ??= _crashIfAnyColumnMatches ?? new Regex(GetPattern(), RegexOptions.IgnoreCase);
 
         foreach (var c in toProcess.Columns.Cast<DataColumn>().Select(c => c.ColumnName))